sensors: rename power package to sensors
