+++ pkg/sensors/accelerator/gpu/gpu.go
@@ -20,7 +20,7 @@ limitations under the License.
 package gpu
 
 import (
-	gpu_source "github.com/sustainable-computing-io/kepler/pkg/power/accelerator/gpu/source"
+	gpu_source "github.com/sustainable-computing-io/kepler/pkg/sensors/accelerator/gpu/source"
 	"k8s.io/klog/v2"
 )
