+++ pkg/sensors/accelerator/gpu/source/type.go
