+++ pkg/sensors/accelerator/qat/source/type.go
@@ -22,14 +22,24 @@ Further understand the device utilization obtained through telemetry by viewing 
 */
 
 type DeviceUtilizationSample struct {
-	SampleCnt   uint64
+	// SampleCnt is a message counter
+	SampleCnt uint64
+	// PciTransCnt is a PCIe Partial Transaction counter
 	PciTransCnt uint64
-	Latency     uint64
-	BwIn        uint64
-	BwOut       uint64
-	CprUtil     uint64
-	DcprUtil    uint64
-	XltUtil     uint64
-	CphUtil     uint64
-	AthUtil     uint64
+	// Latency is the Average Get To Put latency in nanoseconds
+	Latency uint64
+	// BwIn is the PCIe write bandwidth in Mbps
+	BwIn uint64
+	// BwOut is the PCIe read bandwidth in Mbps
+	BwOut uint64
+	// CprUtil is the Compression Slice Utilization On Slice X in percentage execution cycles
+	CprUtil uint64
+	// DcprUtil is the Decompression Slice Utilization On Slice X in percentage execution cycles
+	DcprUtil uint64
+	// XltUtil is the Translator Slice Utilization On Slice X in percentage execution cycles
+	XltUtil uint64
+	// CphUtil is the Cipher Slice Utilization On Slice X in percentage execution cycles
+	CphUtil uint64
+	// AthUtil is the Authentication Slice Utilization On Slice X, percentage execution cycles
+	AthUtil uint64
 }
