+++ pkg/sensors/components/source/apm_xgene_sysfs.go
