+++ pkg/sensors/platform/source/hmc.go
