+++ pkg/sensors/platform/source/redfish.go
