+++ pkg/sensors/platform/source/redfish_test.go
