+++ pkg/sensors/platform/source/redfish_util.go
